requirements.txt
annotated-types==0.6.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
googletrans2==2.3.0
gTTS==2.4.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.2
hyperframe==5.2.0
idna==3.6
pydantic==2.5.2
pydantic-core==2.14.5
python-multipart==0.0.6
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
SpeechRecognition==3.10.0
starlette==0.27.0
typing-extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
deep-translator==1.11.4
openai==1.3.6
python-dotenv
pytest